Move auto-? insertion from codegen to lowering
insert_try_body and all helpers moved from
pass_result_propagation.rs (codegen) to lower/auto_try.rs
(frontend). This is desugaring, not code generation.

Codegen only retains fan-block Try insertion for test functions.
ResultPropagation is now Phase 1 (signature lift) + Phase 2
(wrap tail in Ok) only — no more IR type rewriting.

Pipeline is now clean:
  checker → Result types on effect fn calls
  lowering → auto_try inserts Try nodes (desugaring)
  codegen → signature lift + tail Ok wrap (mechanical)
